requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
python-dotenv
pymupdf
google-genai